+++ native/c/commands/ds.cpp
@@ -697,6 +697,26 @@ int handle_data_set_delete(InvocationContext &context)
   return rc;
 }
 
+int handle_data_set_rename(InvocationContext &context)
+{
+  int rc = 0;
+  string dsn_before = context.get<string>("dsname-before", "");
+  string dsn_after = context.get<string>("dsname-after", "");
+  ZDS zds = {};
+
+  rc = zds_rename_dsn(&zds, dsn_before, dsn_after);
+
+  if (0 != rc)
+  {
+    context.error_stream() << "Error: Could not rename data set: '" << dsn_before << "' rc: '" << rc << "'" << endl;
+    context.error_stream() << " Details: " << zds.diag.e_msg << endl;
+    return RTNCD_FAILURE;
+  }
+  context.output_stream() << "Data set '" << dsn_before << "' renamed to '" << dsn_after << "'" << endl;
+
+  return rc;
+}
+
 int handle_data_set_restore(InvocationContext &context)
 {
   int rc = 0;
@@ -939,6 +959,13 @@ void register_commands(parser::Command &root_command)
   ds_restore_cmd->add_positional_arg(DSN);
   ds_restore_cmd->set_handler(handle_data_set_restore);
   data_set_cmd->add_command(ds_restore_cmd);
+
+  // Rename subcommand
+  auto ds_rename_cmd = command_ptr(new Command("rename", "rename data set"));
+  ds_rename_cmd->add_positional_arg("dsname-before", "data set to rename", ArgType_Single, true);
+  ds_rename_cmd->add_positional_arg("dsname-after", "new data set name", ArgType_Single, true);
+  ds_rename_cmd->set_handler(handle_data_set_rename);
+  data_set_cmd->add_command(ds_rename_cmd);
 
   // Compress subcommand
   auto ds_compress_cmd = command_ptr(new Command("compress", "compress data set"));